fix:  correct the cvat point format with rle point+bbox and fix the schema to points float -> int

+++ src/allocation/domain/pa_dental_segmentation/schemas.py
@@ -23,7 +23,7 @@ class CvatSegmentation(BaseModel):
     confidence: float
     label: str
     type: str
-    points: List[float]
+    points: List[int]
     #mask: List[int]
 
 class CvatSegmentations(BaseModel):
